separates endpoints in files #46
for API versioning in files separate the versions and include them in the main app
